Add files via upload
This was my 2nd Year, Semester 1 Programming project. It is a game designed for the user to guess a random word after receiving various clues about the word. The user has 5 clues and 5 attempts to guess each word. If they get it correct, they will advance to the next word and get a point and if they get it wrong they will move onto the next word and will get no points.